refactor: changes the want the regex works
changes the regex string to be more strict. the older version was
matching on things it should not have been such as `x64-84` or `i386`,
which are not version numbers. so the new regex will only match if it
looks like `0.0.0`, `0.0` or `0`. everything else should be stripped
before checking for the version like hashes or names or whatever (like
`0.1.0-beta`

this involved some rework on the parsing functions and created a new
parse `from_str_with` which allows passing in your own regex to parse a
string. so you could make a regex string that matches `0.1.0-beta` and
will generate a version of `0.1.0` but it will loose that extra bit at
the end. right now this function doesn't have error checking so it will
crash, but thats ok since this is designed for middle use (people using
hte library) not end user, so it may crash on build but it should be ok
as long as you test and don't expose that to the user.

also moves the comments to the correct style so functions should have
some descriptions when using things like `rust-analyzer`