refactor: removes regex
the previous commit changed the regex string, but then i noticed that it
was locking in the version length to 3 digits (since that was the length
of the regex string). i tried to make it expand and do more captures but
it seems that just using a string split is the best thing to do.

so now regex is no longer used. instead it will split the string by a
string which defaults to `.` but can be overwritten using
`from_str_with`. it doesn't work as a pattern, so you will need to
remove any leading or trailing stuff from the version number or it will
fail the parse.